Inference benchmarking metrics
