+++ src/maxdiffusion/configs/base21.yml
@@ -22,6 +22,7 @@ gcs_metrics: True
 save_config_to_gcs: False
 log_period: 100
 
+model_name: 'SDXL-2-1'
 pretrained_model_name_or_path: 'stabilityai/stable-diffusion-2-1'
 revision: 'bf16'
 dtype: 'bfloat16'
